requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django==3.2.6
django-cors-headers==3.7.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
phonenumbers==8.12.28
Pillow==8.3.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
